requirements.stable.txt
requests==2.28.1
selenium==4.3.0
Appium_Python_Client==2.6.0
pytest==7.1.2
pytest-rerunfailures==10.2
allure-pytest==2.9.45
httpx==0.23.0
urllib3==1.26.11
loguru==0.6.0
beautifulsoup4==4.11.1
pytest-progress==1.2.5
pytest-portion==0.1.0
tox==3.25.1
opencv-python==4.6.0.66
imutils==0.5.4
pillow==9.2.0